chore: refactor summary card into separate component
